requirements.txt
and==66.0.2
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
django-heroku==0.3.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
phonenumbers==8.13.20
psycopg2==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
waitress==2.1.2
whitenoise==6.5.0
